Combine identical test routines for pcolor + pcolormesh.
